+++ Intern/rayx-core/src/Beamline/Objects/CircleSource.cpp
@@ -1,7 +1,7 @@
 
 #include "CircleSource.h"
 
-#include "Data/xml.h"
+#include "Rml/xml.h"
 #include "Debug/Debug.h"
 #include "Debug/Instrumentor.h"
 #include "Design/DesignSource.h"
