+++ Intern/rayx-core/src/Beamline/Objects/DipoleSource.cpp
@@ -2,7 +2,7 @@
 
 #include <fstream>
 
-#include "Data/xml.h"
+#include "Rml/xml.h"
 #include "Debug/Debug.h"
 #include "Debug/Instrumentor.h"
 #include "Design/DesignSource.h"
