+++ Intern/rayx-core/src/Beamline/Objects/MatrixSource.cpp
@@ -1,6 +1,6 @@
 #include "MatrixSource.h"
 
-#include "Data/xml.h"
+#include "Rml/xml.h"
 #include "Debug/Debug.h"
 #include "Debug/Instrumentor.h"
 #include "Design/DesignSource.h"
