+++ Intern/rayx-core/src/Beamline/Objects/PixelSource.cpp
@@ -1,6 +1,6 @@
 #include "PixelSource.h"
 
-#include "Data/xml.h"
+#include "Rml/xml.h"
 #include "Debug/Debug.h"
 #include "Debug/Instrumentor.h"
 #include "Design/DesignElement.h"
