+++ Intern/rayx-core/src/Element/Strings.cpp
