+++ Intern/rayx-core/src/Element/Strings.h
