+++ Intern/rayx-core/tests/setupTests.h
@@ -8,7 +8,7 @@
 #include "Beamline/Objects/Objects.h"
 #include "CanonicalizePath.h"
 #include "Core.h"
-#include "Data/Importer.h"
+#include "Rml/Importer.h"
 #include "Debug/Debug.h"
 #include "Design/DesignElement.h"
 #include "Design/DesignSource.h"
